dashboards/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
qrcode==7.4.2
opencv-python==4.8.1.78
face-recognition==1.3.0
Pillow==10.1.0
numpy==1.26.0
openpyxl==3.1.2
reportlab==4.0.4